fix: update authentication checks to use logical OR for role validation

+++ src/app/api/storeInfo/[id]/route.js
@@ -9,7 +9,7 @@ import prisma from '~/app/api/Libs/prisma'
 export const GET = async (request, { params }) => {
   try {
     const { role, userId } = authenticateToken(request) ?? {}
-    if (role !== 'ADMIN' && !userId) return ERROR.FORBIDDEN()
+    if (role !== 'ADMIN' || !userId) return ERROR.FORBIDDEN()
     const { id } = params
     if (!Number(id)) return ERROR.INVALID_FIELDS()
     const payload = await prisma.storeInfo.findUnique({
@@ -30,7 +30,7 @@ export const GET = async (request, { params }) => {
 export const PUT = async (request, { params }) => {
   try {
     const { role, userId } = authenticateToken(request) ?? {}
-    if (role !== 'ADMIN' && !userId) return ERROR.FORBIDDEN()
+    if (role !== 'ADMIN' || !userId) return ERROR.FORBIDDEN()
     const { id } = params
     if (!Number(id)) return ERROR.INVALID_FIELDS()
     const data = await request.json()
@@ -55,7 +55,7 @@ export const PUT = async (request, { params }) => {
 export const PATCH = async (request, { params }) => {
   try {
     const { role, userId } = authenticateToken(request) ?? {}
-    if (role !== 'ADMIN' && !userId) return ERROR.FORBIDDEN()
+    if (role !== 'ADMIN' || !userId) return ERROR.FORBIDDEN()
     const { id } = params
     if (!Number(id)) return ERROR.INVALID_FIELDS()
     const data = await request.json()
@@ -78,7 +78,7 @@ export const PATCH = async (request, { params }) => {
 export const DELETE = async (request, { params }) => {
   try {
     const { role, userId } = authenticateToken(request) ?? {}
-    if (role !== 'ADMIN' && !userId) return ERROR.FORBIDDEN()
+    if (role !== 'ADMIN' || !userId) return ERROR.FORBIDDEN()
     const { id } = params
     if (!Number(id)) return ERROR.INVALID_FIELDS()
     const payload = await prisma.storeInfo.delete({
